Fixed several issues with logger unit tests.

+++ SitecoreSpark.SPRK.Tests/SparkLoggerTests.cs
@@ -10,7 +10,7 @@ namespace SitecoreSpark.SPRK.Tests
     public class SparkLoggerTests
     {
         internal string _logFolderPath = "./TESTLOGS";
-        internal string _logFilePrefix = "SPRK.TESTLOG.";
+        internal string _logFilePrefix = "SPRK.TESTLOG";
         internal int _logBufferSize = 1000;
 
         private SparkLogger SetupLogger()
@@ -21,11 +21,14 @@ namespace SitecoreSpark.SPRK.Tests
 
         private string GetTestFilePath()
         {
-            return Path.Combine(_logFolderPath, ($"{this._logFilePrefix}{DateTime.Now.ToString("yyyyMMdd")}.txt"));
+            return Path.Combine(_logFolderPath, ($"{this._logFilePrefix}.{DateTime.Now.ToString("yyyyMMdd")}.txt"));
         }
 
         private void CleanUpTestLogFiles()
         {
+            if (!Directory.Exists(this._logFolderPath))
+                return;
+
             foreach (var filePath in Directory.GetFiles(this._logFolderPath))
             {
                 File.Delete(filePath);
@@ -76,6 +79,8 @@ namespace SitecoreSpark.SPRK.Tests
             logger.CloseLog(testKey);
 
             Assert.IsFalse(Directory.Exists(this._logFolderPath));
+
+            this.CleanUpTestLogFiles();
         }
 
         [TestMethod]
